requirements.txt
awsebcli==3.7.2
blessed==1.9.5
botocore==1.3.22
cement==2.4.0
colorama==0.3.3
docker-py==1.1.0
dockerpty==0.3.4
docopt==0.6.2
docutils==0.12
Flask==0.10.1
Flask-MySQL==1.3
Flask-MySQLdb==0.2.0
httplib2==0.9.2
itsdangerous==0.24
Jinja2==2.8
jmespath==0.9.0
MarkupSafe==0.23
MySQL-python==1.2.5
mysqlclient==1.3.7
pathspec==0.3.3
python-dateutil==2.4.2
pytz==2015.7
PyYAML==3.11
requests==2.6.2
six==1.10.0
texttable==0.8.4
twilio==5.2.0
wcwidth==0.1.6
websocket-client==0.35.0
Werkzeug==0.11.3
wheel==0.24.0
